+++ etna/analysis/feature_selection/mrmr_selection.py
@@ -35,6 +35,7 @@ def mrmr(
     regressors: pd.DataFrame,
     top_k: int,
     fast_redundancy: bool = False,
+    drop_zero: bool = False,
     relevance_aggregation_mode: str = AggregationMode.mean,
     redundancy_aggregation_mode: str = AggregationMode.mean,
     atol: float = 1e-10,
@@ -59,6 +60,10 @@ def mrmr(
     fast_redundancy:
         * True: compute redundancy only inside the the segments, time complexity :math:`O(top\_k * n\_segments * n\_features * history\_len)`
         * False: compute redundancy for all the pairs of segments, time complexity :math:`O(top\_k * n\_segments^2 * n\_features * history\_len)`
+    drop_zero:
+        * True: use only features with relevance > 0 in calculations, if their number is less than ``top_k``
+            randomly selects features with zero relevance so that the total number of selected features is ``top_k``
+        * False: use all features in calculations
     relevance_aggregation_mode:
         the method for relevance values per-segment aggregation
     redundancy_aggregation_mode:
@@ -88,16 +93,26 @@ def mrmr(
     relevance = relevance_table.apply(relevance_aggregation_fn).fillna(0)
 
     all_features = relevance.index.to_list()
+
+    if top_k >= len(all_features):
+        return all_features.copy()
+
     segments = set(regressors.columns.get_level_values("segment"))
     selected_features: List[str] = []
     not_selected_features = all_features.copy()
 
-    redundancy_table = pd.DataFrame(np.inf, index=all_features, columns=all_features)
-    top_k = min(top_k, len(all_features))
+    if drop_zero:
+        not_relevant_features = list(filter(lambda feature: not relevance.loc[feature] == 0, not_selected_features))
+        relevant_features = list(set(all_features) - set(not_relevant_features))
+        if top_k >= len(relevant_features):
+            return relevant_features + not_relevant_features[: (top_k - len(relevant_features))]
+        not_selected_features = relevant_features
+
+    redundancy_table = pd.DataFrame(1, index=all_features, columns=all_features)
 
     for i in range(top_k):
         score_numerator = relevance.loc[not_selected_features]
-        score_denominator = pd.Series(1, index=not_selected_features)
+        score_denominator = pd.Series(0, index=not_selected_features)
         if i > 0:
             last_selected_feature = selected_features[-1]
             last_selected_regressor = regressors.loc[pd.IndexSlice[:], pd.IndexSlice[:, last_selected_feature]]
@@ -124,14 +139,13 @@ def mrmr(
             redundancy_table.loc[not_selected_features, last_selected_feature] = (
                 segment_redundancy.agg(redundancy_aggregation_fn)
                 .clip(atol)
-                .fillna(np.inf)
+                .fillna(1)
                 .loc[not_selected_features]
                 .values.squeeze()
             )
 
-            score_denominator = redundancy_table.loc[not_selected_features, selected_features].mean(axis=1)
-            score_denominator[np.isclose(score_denominator, 1, atol=atol)] = np.inf
-        score = score_numerator / score_denominator
+            score_denominator = redundancy_table.loc[not_selected_features, selected_features].max(axis=1)
+        score = score_numerator * (1 - score_denominator)
         best_feature = score.index[score.argmax()]
         selected_features.append(best_feature)
         not_selected_features.remove(best_feature)