+++ etna/transforms/feature_selection/feature_importance.py
@@ -172,6 +172,7 @@ class MRMRFeatureSelectionTransform(BaseFeatureSelectionTransform):
         top_k: int,
         features_to_use: Union[List[str], Literal["all"]] = "all",
         fast_redundancy: bool = False,
+        drop_zero: bool = False,
         relevance_aggregation_mode: str = AggregationMode.mean,
         redundancy_aggregation_mode: str = AggregationMode.mean,
         atol: float = 1e-10,
@@ -190,6 +191,10 @@ class MRMRFeatureSelectionTransform(BaseFeatureSelectionTransform):
         features_to_use:
             columns of the dataset to select from
             if "all" value is given, all columns are used
+        drop_zero:
+            * True: use only features with relevance > 0 in calculations, if their number is less than ``top_k``,
+              randomly selects features with zero relevance so that the total number of selected features is ``top_k``
+            * False: use all features in calculations
         fast_redundancy:
             * True: compute redundancy only inside the segments, time complexity :math:`O(top\_k \\cdot n\_segments \\cdot n\_features \\cdot history\_len)`
             * False: compute redundancy for all the pairs of segments, time complexity :math:`O(top\_k \\cdot n\_segments^2 \\cdot n\_features \\cdot history\_len)`
@@ -209,6 +214,7 @@ class MRMRFeatureSelectionTransform(BaseFeatureSelectionTransform):
         self.relevance_table = relevance_table
         self.top_k = top_k
         self.fast_redundancy = fast_redundancy
+        self.drop_zero = drop_zero
         self.relevance_aggregation_mode = relevance_aggregation_mode
         self.redundancy_aggregation_mode = redundancy_aggregation_mode
         self.atol = atol
@@ -236,12 +242,16 @@ class MRMRFeatureSelectionTransform(BaseFeatureSelectionTransform):
         relevance_table = self.relevance_table(df_target, df_features, **self.relevance_params)
 
         if not self.relevance_table.greater_is_better:
-            relevance_table *= -1
+            min_relevance = relevance_table.values.min()
+            max_relevance = relevance_table.values.max()
+            relevance_table = max_relevance + min_relevance - relevance_table
+
         self.selected_features = mrmr(
             relevance_table=relevance_table,
             regressors=df_features,
             top_k=self.top_k,
             fast_redundancy=self.fast_redundancy,
+            drop_zero=self.drop_zero,
             relevance_aggregation_mode=self.relevance_aggregation_mode,
             redundancy_aggregation_mode=self.redundancy_aggregation_mode,
             atol=self.atol,